Moved IChatHistoryReducer to Abstractions package
