+++ dotnet/samples/GettingStartedWithProcesses/Step04/KernelExtensions.cs
@@ -1,8 +1,9 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
+
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.SemanticKernel;
 using Microsoft.SemanticKernel.Agents;
-using Microsoft.SemanticKernel.Agents.History;
+using Microsoft.SemanticKernel.ChatCompletion;
 
 namespace Step04;
