+++ dotnet/src/SemanticKernel.Abstractions/AI/ChatCompletion/IChatHistoryReducer.cs
@@ -1,0 +1,23 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+
+using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Microsoft.SemanticKernel.ChatCompletion;
+
+/// <summary>
+/// Interface for reducing the chat history.
+/// </summary>
+[Experimental("SKEXP0001")]
+public interface IChatHistoryReducer
+{
+    /// <summary>
+    /// Reduces the <see cref="ChatHistory"/>.
+    /// </summary>
+    /// <param name="history">Instance of <see cref="ChatHistory"/>to be reduced.</param>
+    /// <param name="cancellationToken">The <see cref="CancellationToken"/> to monitor for cancellation requests. The default is <see cref="CancellationToken.None"/>.</param>
+    /// <returns>The reduced history or <see langword="null"/> if no reduction has occurred.</returns>
+    Task<IEnumerable<ChatMessageContent>?> ReduceAsync(IReadOnlyList<ChatMessageContent> history, CancellationToken cancellationToken = default);
+}
